Seeding Service

+++ SalesWebMVC/Models/Department.cs
@@ -27,7 +27,7 @@ namespace SalesWebMVC.Models
             Sellers.Add(seller); 
         }
 
-        public decimal TotalSales(DateTime initial, DateTime final)
+        public double TotalSales(DateTime initial, DateTime final)
         {
             return Sellers.Sum(seller => seller.TotalSales(initial, final));
         }
