+++ SalesWebMVC/Models/SalesRecord.cs
@@ -7,7 +7,7 @@ namespace SalesWebMVC.Models
     {
         public int SalesRecordId { get; set; }
         public DateTime Date { get; set; }
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public SaleStatus Status { get; set; }
         public Seller Seller { get; set; }
 
@@ -15,7 +15,7 @@ namespace SalesWebMVC.Models
         {
         }
 
-        public SalesRecord(int salesRecordId, DateTime date, decimal amount, SaleStatus status, Seller seller)
+        public SalesRecord(int salesRecordId, DateTime date, double amount, SaleStatus status, Seller seller)
         {
             SalesRecordId = salesRecordId;
             Date = date;
