+++ SalesWebMVC/Models/Seller.cs
@@ -11,7 +11,7 @@ namespace SalesWebMVC.Models
         public string Name { get; set; }
         public string Email { get; set; }
         public DateTime BirthDate { get; set; }
-        public decimal BaseSalary { get; set; }
+        public double BaseSalary { get; set; }
         public Department Department { get; set; }
         public ICollection<SalesRecord> Sales { get; set; } = new List<SalesRecord>();
 
@@ -19,7 +19,7 @@ namespace SalesWebMVC.Models
         {
         }
 
-        public Seller(int sellerId, string name, string email, DateTime birthDate, decimal baseSalary, Department department)
+        public Seller(int sellerId, string name, string email, DateTime birthDate, double baseSalary, Department department)
         {
             SellerId = sellerId;
             Name = name;
@@ -39,7 +39,7 @@ namespace SalesWebMVC.Models
             Sales.Remove(sr);
         }
 
-        public decimal TotalSales(DateTime initial, DateTime final)
+        public double TotalSales(DateTime initial, DateTime final)
         {
             return Sales.Where(sr => sr.Date >= initial && sr.Date <= final).Sum(sr => sr.Amount);
         }
